services2/preprocess/requirements.txt
flask==2.2.3
pandas==1.5.3
python-dotenv==1.0.0
Pillow==9.4.0
torch==2.0.0
scikit-learn==1.2.2
opencv-python==4.7.0.68
numpy==1.24.2
transformers==4.27.4
minio==7.1.14
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
celery==5.2.7
pymilvus==2.2.6
google
grpcio
protobuf